[ECP-9320] Updating constants
